+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0-alpha] - 2024-08-28
+
 ### Added
 
 - `Socket.Available` implementation
@@ -109,8 +111,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configurable SOCKS commands.
 - Configurable allow and block lists.
 
-[Unreleased]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.1-alpha...HEAD
-[0.3.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.0-alpha...v0.3.1-alpha
+[Unreleased]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.4.0-alpha...HEAD
+[0.4.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.1-alpha...v0.4.0-alpha
+[0.3.1-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.3.0-alpha...v0.3.1-alpha
 [0.3.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.2.0-alpha...v0.3.0-alpha
 [0.2.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/compare/v0.1.0-alpha...v0.2.0-alpha
 [0.1.0-alpha]: https://github.com/TSRBerry/RyuSOCKS/releases/tag/v0.1.0-alpha